Big update ! About half of the job done.
VC armybook latexed from beginning to lords section, not included.
Numerous updates in the layout code.
AB_layout.tex is still a HUGE MESS, so if you wanna help me, better that
we discuss it before, because half of AB_layout.tex is due to be
deleted, and I have things in mind for the other half !